+++ tests/value-init.test.ts
@@ -18,26 +18,26 @@ describe("value-type initializers (with symtab)", () => {
   test("init on a small loadable struct returns an owned ValueInstance", () => {
     const v = structType("fixturesyms.Point").init(5);
     expect(v.$owned).toBe(true);
-    expect(v.$fields).toEqual({ x: 5 });
+    expect(v.$fields).toEqual({ x: int64(5) });
     v.$dispose();
   });
 
   test("init marshals a String arg and adopts a non-POD return", () => {
     const v = structType("fixturesyms.Person").init("Ada", 36);
-    expect(v.$fields).toEqual({ name: "Ada", age: 36 });
+    expect(v.$fields).toEqual({ name: "Ada", age: int64(36) });
     v.$dispose();
   });
 
   test("init adopts a large struct returned indirectly", () => {
     const v = structType("fixturesyms.BigStruct").init(1, 2, 3, 4, 5);
-    expect(v.$fields).toEqual({ a: 1, b: 2, c: 3, d: 4, e: 5 });
+    expect(v.$fields).toEqual({ a: int64(1), b: int64(2), c: int64(3), d: int64(4), e: int64(5) });
     v.$dispose();
   });
 
   test("a bound initializer is reusable across calls", () => {
     const make = structType("fixturesyms.Point").initializer();
-    expect(make.call(1).$fields).toEqual({ x: 1 });
-    expect(make.call(2).$fields).toEqual({ x: 2 });
+    expect(make.call(1).$fields).toEqual({ x: int64(1) });
+    expect(make.call(2).$fields).toEqual({ x: int64(2) });
   });
 
   test("throws on an argument-count mismatch", () => {
